Make all settings int

+++ BeerSpec_GUI/SettingsClass.h
@@ -5,12 +5,12 @@ private:
 	int m_LEDR;
 	int m_LEDG;
 	int m_LEDB;
-	float m_GainExtR;
-	float m_GainExtG;
-	float m_GainExtB;
-	float m_GainScaR;
-	float m_GainScaG;
-	float m_GainScaB;
+	int m_GainExtR;
+	int m_GainExtG;
+	int m_GainExtB;
+	int m_GainScaR;
+	int m_GainScaG;
+	int m_GainScaB;
 	int m_NumSamples;
 
 public:
